refactor: Update obtenerCuestionariosPorEstado method in CuestionariosService
This commit refactors the `obtenerCuestionariosPorEstado` method in the `CuestionariosService` class. It updates the method to include additional nested selections in the Prisma query, allowing for more detailed information to be retrieved. This change enhances the functionality of the service by providing a richer set of data when retrieving cuestionarios by estado.