refactor: Set default value for 'cant_leido' field in Libro model
This commit sets the default value of the 'cant_leido' field in the Libro model to 0. This change ensures that the 'cant_leido' field is always initialized with a value, even if it is not explicitly provided. This improves the consistency and reliability of the data stored in the 'cant_leido' field.