refactor: Add descripcion field to GeneroLiterario model
This commit adds the 'descripcion' field to the 'GeneroLiterario' model in the Prisma schema. The 'descripcion' field is of type 'String' and is used to store the description of the literary genre. This change enhances the data model by providing additional information about each genre, improving the overall functionality and user experience of the application.