feat: Add 'descripcion' field to GeneroLiterario entity
This commit adds the 'descripcion' field to the 'GeneroLiterario' entity. The 'descripcion' field is used to store the description of the literary genre and is required to be between 3 and 500 characters long. This change enhances the data model by providing additional information about each genre, improving the overall functionality and user experience of the application.