feat: Add endpoint to search books by literary genre
This commit adds a new endpoint to the LibrosController that allows users to search for books by literary genre. The endpoint is implemented in the `obtenerLibrosPorGeneroLiterario` method and accepts a query parameter `genero_literario` to specify the genre. This feature enhances the functionality of the application by providing a convenient way for users to discover books based on their preferred genre.