feat: Include grado_usuario in obtenerUsuarioForProfile response
This commit modifies the `obtenerUsuarioForProfile` method in the `UsuariosService` to include the `grado_usuario` information in the response. By selecting the `nom_grado` field from the `grados` table, the method now provides the user's degree information along with their avatar and profile photo. This enhancement improves the completeness of the user profile data and enhances the functionality of the application.