refactor: Update CrearCuestionarioDto to handle null values for fecha_asignado and fecha_entrega
This commit updates the CrearCuestionarioDto to handle null values for the fecha_asignado and fecha_entrega properties. It uses a conditional transform to convert the input values to Date objects if they are not null. This change improves the flexibility and usability of the application by allowing the fecha_asignado and fecha_entrega fields to be optional when creating a cuestionario.