refactor: Update Avatar model to establish a one-to-many relationship with AvatarUsuario
This commit modifies the Avatar model in the prisma/schema.prisma file to change the relationship with AvatarUsuario from one-to-one to one-to-many. The avatar_usuario field is now an array of AvatarUsuario objects, allowing multiple AvatarUsuario instances to be associated with a single Avatar. This change improves the flexibility and functionality of the application when working with avatars and users.