refactor: Update CrearAvatarUsuarioDto to handle optional id_avatar
This commit updates the CrearAvatarUsuarioDto class in the avatar-usuario/dto/avatar-usuario.dto.ts file. It adds the @Transform decorator to the id_avatar property, allowing it to handle optional values. This change improves the flexibility of the DTO when creating avatar usuarios by allowing the id_avatar field to be optional.