refactor: Update UsuariosService to include avatar name in query
This commit updates the UsuariosService class in the usuarios/usuarios.service.ts file. It modifies the include option for the avatar_usuario property to also include the name of the avatar. This change enhances the functionality of the service by providing the name of the avatar along with the user data, allowing for easier identification and display of the avatar associated with each user.