refactor: Add foto_perfil and avatar_usuario properties to RegistrarseService
This commit adds the foto_perfil and avatar_usuario properties to the RegistrarseService class in the registrarse/registrarse.service.ts file. These properties are created as empty objects when registering a new user. This change enhances the functionality of the service by providing the necessary structure for storing and associating profile photos and avatars with user registrations.