refactor: Add use_avatar property to Usuario model
This commit adds the use_avatar property to the Usuario model in the prisma/schema.prisma file. The use_avatar property is a boolean that indicates whether the user has uploaded a profile photo or not. This change enhances the functionality of the application by providing a way to track and manage user profile photos.