refactor: Update AvatarUsuarioController to update user's use_avatar property
This commit updates the AvatarUsuarioController in the avatar-usuario/avatar-usuario.controller.ts file. It adds logic to update the use_avatar property of the user when updating the avatar. This change enhances the functionality of the controller by ensuring that the use_avatar property is correctly updated when the avatar is changed.