refactor: Update FotoPerfilController constructor to include UsuariosService
This commit updates the constructor of the FotoPerfilController in the foto-perfil/foto-perfil.controller.ts file. It adds the UsuariosService as a parameter to the constructor and assigns it to the private readonly usuariosService property. This change enhances the functionality of the controller by allowing it to access and use the UsuariosService for updating user information related to profile photos.